@@ -4,6 +4,7 @@
   "version": "0.0.0",
   "author": "typhonrt",
   "bugs": "https://github.com/typhonjs-node-rollup/issues/issues",
+  "type": "module",
   "dependencies": {
     "@rollup/plugin-json": "^4.0.0"
   },
@@ -22,7 +23,7 @@
     "rollup": "^2.x"
   },
   "engines": {
-    "node": ">=8.0.0"
+    "node": ">=14.0.0"
   },
   "main": "src/loader/PluginLoader.js",
   "files": [
